src/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-debug-toolbar==3.1.1
pytz==2020.1
sqlparse==0.4.1
djangorestframework==3.11.0
